+++ package/src/renderer/processors/Transform.ts
@@ -1,8 +1,6 @@
 import type { DrawingContext } from "../DrawingContext";
-import type { SkMatrix, Vector } from "../../skia/types";
-
-import { processTransform2d } from "./math";
-import type { Transforms2d } from "./math";
+import type { SkMatrix, Vector, Transforms2d } from "../../skia/types";
+import { processTransform } from "../../skia/types";
 
 export interface TransformProps {
   transform?: Transforms2d;
@@ -10,7 +8,7 @@ export interface TransformProps {
   matrix?: SkMatrix;
 }
 
-export const processTransform = (
+export const processCanvasTransform = (
   { canvas, Skia }: DrawingContext,
   { transform, origin, matrix }: TransformProps
 ) => {
@@ -18,7 +16,7 @@ export const processTransform = (
     if (matrix) {
       canvas.concat(matrix);
     } else {
-      const m3 = processTransform2d(
+      const m3 = processTransform(
         Skia.Matrix(),
         origin ? transformOrigin(origin, transform) : transform
       );
@@ -32,7 +30,7 @@ export const localMatrix = (
   { transform, origin }: TransformProps
 ) => {
   if (transform) {
-    return processTransform2d(
+    return processTransform(
       m,
       origin ? transformOrigin(origin, transform) : transform
     );
