+++ package/src/renderer/processors/math/index.ts
@@ -1,4 +1,3 @@
 export * from "./Coordinates";
-export * from "./Matrix3";
 export * from "./Math";
 export * from "./Transforms";
