+++ package/src/skia/core/Matrix.ts
@@ -1,0 +1,6 @@
+import { Skia } from "../Skia";
+import type { Transforms2d } from "../types";
+import { processTransform } from "../types";
+
+export const processTransform2d = (transforms: Transforms2d) =>
+  processTransform(Skia.Matrix(), transforms);
